+++ preferences.py
@@ -18,40 +18,63 @@ def get_addon_prefs():
     return (addon_prefs)
 
 
-class myaddonPrefs(bpy.types.AddonPreferences):
-    ## can be just __name__ if prefs are in the __init__ mainfile
-    # Else need the splitext '__name__ = addonname.subfile' (or use a static name)
-    bl_idname = __name__.split('.')[0] # or with: os.path.splitext(__name__)[0]
+class SWD_OT_check_ffmpeg(bpy.types.Operator):
+    """check if ffmpeg is in path"""
+    bl_idname = "swd.check_ffmpeg"
+    bl_label = "Check ffmpeg in system path"
+    bl_options = {'REGISTER', 'INTERNAL'}
+    
+    def invoke(self, context, event):
+        import  shutil
+        self.ok = shutil.which('ffmpeg')
+        return context.window_manager.invoke_props_dialog(self, width=250)
+    
+    def draw(self, context):
+        layout = self.layout
+        if self.ok:
+            layout.label(text='Ok ! ffmpeg is in system PATH', icon='INFO')
+        else:
+            layout.label(text='ffmeg is not in system PATH', icon='CANCEL')
 
-    # some_bool_prop to display in the addon pref
-    super_special_option : bpy.props.BoolProperty(
-        name='Use super special option',
-        description="This checkbox toggle the use of the super special options",
-        default=False)
+    def execute(self, context):
+        return {'FINISHED'}
+
+class SWD_sound_waveform_display_addonpref(bpy.types.AddonPreferences):
+    bl_idname = __package__
+    # bl_idname = __name__.split('.')[0] # or with: os.path.splitext(__name__)[0]
+
+    path_to_ffmpeg : StringProperty(
+        name="Path to ffmpeg binary",
+        subtype='FILE_PATH')
 
     def draw(self, context):
-            layout = self.layout
+        layout = self.layout
+        layout.use_property_split = True
+        # layout.use_property_decorate = False
+        box = layout.box()
+        col = box.column()
+        # col.label(text="This addon use ffmpeg to generate the waveform (need a recent version)")
+        
+        row = col.row()
+        row.label(text="This functionallity need a recent ffmpeg binary")
+        row.operator('wm.url_open', text='ffmpeg download page', icon='URL').url = 'https://www.ffmpeg.org/download.html'
+        
+        row = col.row()
+        row.label(text="Leave field empty if ffmpeg is in system PATH")
+        row.operator('swd.check_ffmpeg', text='Check if ffmpeg in PATH', icon='PLUGIN')
+        
+        # col.label(text="May not work if space are in path.")
+        box.prop(self, "path_to_ffmpeg")
 
-            ## some 2.80 UI options
-            # layout.use_property_split = True
-            # flow = layout.grid_flow(row_major=True, columns=0, even_columns=True, even_rows=False, align=False)
-            # layout = flow.column()
-
-            layout.label(text='Some text')
-
-            # display the bool prop
-            layout.prop(self, "super_special_option")
-
-            # draw something only if a prop evaluate True
-            if self.super_special_option:
-                layout.label(text="/!\ Carefull, the super special option is especially powerfull")
-                layout.label(text="    and with great power... well you know !")
+        ## Make an auto-install (at least for windows user), maybe store bin on a public repo...
+        
 
 
 ### --- REGISTER ---
 
 classes=(
-myaddonPrefs,
+SWD_OT_check_ffmpeg,
+SWD_sound_waveform_display_addonpref,
 )
 
 def register():